Renaming folder name for the standard 'src'
