+++ example/explorer.php
@@ -1,6 +1,6 @@
 <pre><?php
 
-require_once('../lib/Blockchain.php');
+require_once('../src/Blockchain.php');
 
 $api_code = null;
 if(file_exists('code.txt')) {
